requirements.txt
Django==2.1.7
mysql==0.0.2
mysqlclient==1.4.2.post1
PyMySQL==0.9.3
pytz==2018.9
xadmin==2.0.1
xlrd==1.2.0
xlwt==1.3.0
pillow==6.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.11
 django-pure-pagination==0.3.0